Add support for URL param 'fields'
The 'fields' URL param controls what fields are returned from the search
service. We add special logic to ensure that if 'fields' is set, it
lists the field 'type', since we always need to know the hit type in
order to be able to determine how to decode it. Without it, all hits
would be assumed to be of type 'asset', which is incorrect.

Issue #12