Avoid excessive copying of HitSubset
A new instance of HitSubset is created each time the Subset method is
called, causing excessive copying of fields when Subset is called
multiple times for the same Hit.

Example use-case where this problem arises:

    func LocalizedTitle(hit search.Hit, locale string) string {
            sub := hit.Subset()
            switch locale {
            case "sv_SE":
                    return sub.TitleSv
            case "...":
                    // ...
            }
    }

    func LocalizedDescription(hit search.Hit, locale string) string {
            sub := hit.Subset()
            switch locale {
            case "sv_SE":
                    return sub.DescriptionSv
            case "...":
                    // ...
            }
    }

Have Subset return a *HitSubset instead whose value is cached in the
respective underlying type.